refactoring: Vec2R(x=..., y=...) --> Vec2R(..., ...)
Earlier, when Pyrsistent was used for Vec2R, only the keyword argument
constructor was supported. After Vec2R was changed to be a custom tuple
subclass, the positional argument constructor became possible. With this
commit, the old keyword argument constructor, which is still supported,
is replaced with the positional argument variant throughout the code
base.

This should make the code more compact and readable.