ngspice.py: select_backend and find_library simplified
If an OSError is raised by ctypes.CDLL, the OSError should be
shown to the user. The previous select_backend + find_library code
caught this OSError and did not register it as cause of subsequent
exceptions (raise ... from exc). In this particular situation, the
OSError was not "cannot open shared object file: No such file or
directory" but a "undefined symbol" error due to an improperly built
.so file.