webui: race condition eliminated
Problem: Occasionally, upon loading the webui, there were no views
to choose from in the view selector drop-down menus.

The following occasional order of events caused this problem:

1. loadLayout is called, but window.ordecClient.resultViewers is still
   an empty array.
2. As part of loadLayout, ResultViewers are created, but they do not
   know about the available views yet, as window.ordecClient.views is
   also still empty.
3. ordecClient.connect() is called.
4. 'msglist' message is received, but since the ResultViewers have not
   yet been inserted into ordecClient.result, this list is not sent
   to any of the ResultViewers.
5. The 'stateChanged' event from layout is triggered, causing the
   ResultViewers to be inserted into ordecClient.result.

If events 4 and 5 were swapped, the problem did not occur.

In the fixed code, the problem should not happen anymore. Firstly,
the OrdecClient is only created once the layout is loaded. We don't
wait for the first 'stateChanged' anymore. Apparently, the model is
updated right away after the loadLayout call.

This is also a good occasion to further improve the webui JS code.